+++ mullvad-daemon/src/device/api.rs
@@ -1,7 +1,8 @@
 use std::pin::Pin;
 
+use chrono::{DateTime, Utc};
 use futures::{future::FusedFuture, Future};
-use mullvad_types::{device::Device, wireguard::WireguardData};
+use mullvad_types::{account::VoucherSubmission, device::Device, wireguard::WireguardData};
 
 use super::{Error, PrivateAccountAndDevice, ResponseTx};
 
@@ -34,11 +35,27 @@ impl CurrentApiCall {
         self.current_call = Some(Call::Validation(validation));
     }
 
+    pub fn set_expiry_check(&mut self, expiry_call: ApiCall<DateTime<Utc>>) {
+        self.current_call = Some(Call::ExpiryCheck(expiry_call));
+    }
+
+    pub fn set_voucher_submission(
+        &mut self,
+        voucher_call: ApiCall<VoucherSubmission>,
+        tx: ResponseTx<VoucherSubmission>,
+    ) {
+        self.current_call = Some(Call::VoucherSubmission(voucher_call, Some(tx)));
+    }
+
     pub fn is_validating(&self) -> bool {
         matches!(
             &self.current_call,
             Some(Call::Validation(_)) | Some(Call::OneshotKeyRotation(_))
         )
+    }
+
+    pub fn is_checking_expiry(&self) -> bool {
+        matches!(&self.current_call, Some(Call::ExpiryCheck(_)))
     }
 
     pub fn is_running_timed_totation(&self) -> bool {
@@ -88,6 +105,11 @@ enum Call {
     TimerKeyRotation(ApiCall<WireguardData>),
     OneshotKeyRotation(ApiCall<WireguardData>),
     Validation(ApiCall<Device>),
+    VoucherSubmission(
+        ApiCall<VoucherSubmission>,
+        Option<ResponseTx<VoucherSubmission>>,
+    ),
+    ExpiryCheck(ApiCall<DateTime<Utc>>),
 }
 
 impl futures::Future for Call {
@@ -110,6 +132,17 @@ impl futures::Future for Call {
                 Pin::new(call).poll(cx).map(ApiResult::Rotation)
             }
             Validation(call) => Pin::new(call).poll(cx).map(ApiResult::Validation),
+            VoucherSubmission(call, tx) => {
+                if let std::task::Poll::Ready(response) = Pin::new(call).poll(cx) {
+                    std::task::Poll::Ready(ApiResult::VoucherSubmission(
+                        response,
+                        tx.take().unwrap(),
+                    ))
+                } else {
+                    std::task::Poll::Pending
+                }
+            }
+            ExpiryCheck(call) => Pin::new(call).poll(cx).map(ApiResult::ExpiryCheck),
         }
     }
 }
@@ -118,4 +151,9 @@ pub(crate) enum ApiResult {
     Login(Result<PrivateAccountAndDevice, Error>, ResponseTx<()>),
     Rotation(Result<WireguardData, Error>),
     Validation(Result<Device, Error>),
+    VoucherSubmission(
+        Result<VoucherSubmission, Error>,
+        ResponseTx<VoucherSubmission>,
+    ),
+    ExpiryCheck(Result<DateTime<Utc>, Error>),
 }